Improved challenge solution

+++ rust_linkedin_learning_tutorials/module_2_code_samples/module_2_challenge/src/main.rs
@@ -4,7 +4,7 @@ fn main() {
     let c: f32 = 120.0; 
 
     /* My Code*/
-    let average: f64 = (a as f64 + b as f64 + c as f64) / 3.0; 
+    let average: f64 = (a as f64 + b + c as f64) / 3.0; 
 
     assert_eq!(average, 45.1);
     println!("Test passed!");
